Finalizando TP02...

+++ src/main/java/gerenciador/AlteraProdutoServlet.java
@@ -1,3 +1,4 @@
+// Dupla: Adriano Junior de Souza Almeida e Arthur Lanzilotti Farjanes
 package gerenciador;
 
 import java.io.IOException;
@@ -10,39 +11,34 @@ import javax.servlet.http.HttpServletResponse;
 @WebServlet("/alteraProduto")
 public class AlteraProdutoServlet extends HttpServlet {
 
-	    private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	    protected void doPost(HttpServletRequest request, HttpServletResponse response) 
-	            throws ServletException, IOException {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) 
+            throws ServletException, IOException {
 
-	        System.out.println("Alterando produto v2.0");
+        System.out.println("Alterando produto v2.0");
 
-	        // 1. Captura os parâmetros do formulário
-	        String paramId = request.getParameter("id");
-	        String nome = request.getParameter("nome");
-	        String descricao = request.getParameter("descricao");
-	        String paramUnidade = request.getParameter("unidadeCompra");
-	        String paramQtd = request.getParameter("qtdPrevistoMes");
-	        String paramPreco = request.getParameter("precoMaxComprado");
+        String paramId = request.getParameter("id");
+        String nome = request.getParameter("nome");
+        String descricao = request.getParameter("descricao");
+        String paramUnidade = request.getParameter("unidadeCompra");
+        String paramQtd = request.getParameter("qtdPrevistoMes");
+        String paramPreco = request.getParameter("precoMaxComprado");
 
-	        // 2. Conversão de tipos
-	        Integer id = Integer.valueOf(paramId);
-	        int unidade = Integer.parseInt(paramUnidade);
-	        int qtd = Integer.parseInt(paramQtd);
-	        double preco = Double.parseDouble(paramPreco);
+        Integer id = Integer.valueOf(paramId);
+        int unidade = Integer.parseInt(paramUnidade);
+        double qtd = Double.parseDouble(paramQtd); 
+        double preco = Double.parseDouble(paramPreco);
 
-	        // 3. Busca o produto original no "Banco"
-	        Banco banco = new Banco();
-	        Produto produto = banco.buscaProdutoPelaId(id);
-	        
-	        // 4. Atualiza os dados do objeto
-	        produto.setNome(nome);
-	        produto.setDescricao(descricao);
-	        produto.setUnidadeCompra(unidade);
-	        produto.setQtdPrevistoMes(qtd);
-	        produto.setPrecoMaxComprado(preco);
+        Banco banco = new Banco();
+        Produto produto = banco.buscaProdutoPelaId(id);
+        
+        produto.setNome(nome);
+        produto.setDescricao(descricao);
+        produto.setUnidadeCompra(unidade);
+        produto.setQtdPrevistoMes(qtd);
+        produto.setPrecoMaxComprado(preco);
 
-	        // 5. Redireciona para a lista atualizada
-	        response.sendRedirect("listaProdutos");
-	    }
+        response.sendRedirect("listaProdutos");
+    }
 }
